move the scripts
