--- Assets/Dev/Si Wai/PlayerOrientation.cs
@@ -1,29 +1,0 @@
-using UnityEngine;
-
-public class PlayerOrientation : MonoBehaviour
-{
-    public float sensitivityX = 200f;
-
-    public Rigidbody playerObject;
-
-    private float _yRotation;
-
-    private void Start()
-    {
-        Cursor.lockState = CursorLockMode.Locked;
-        Cursor.visible = false;
-    }
-
-    private void Update()
-    {
-        var trans = this.transform;
-        
-        this._yRotation += Input.GetAxis("Mouse X") * Time.deltaTime * this.sensitivityX;
-        trans.localRotation = Quaternion.Euler(0f, this._yRotation, 0f);
-        
-        trans.position = this.playerObject.transform.position;
-        trans.Translate(Vector3.forward);
-        Debug.DrawRay(trans.position, trans.forward, Color.red, 0f, false);
-        
-    }
-}
