+++ Assets/Global/Scripts/PlayerObjScript.cs.meta
